complete cartitem component

+++ backend/cart/models.py
@@ -6,8 +6,8 @@ class Cart(models.Model):
         db_table = 'cart'
     
     id = models.AutoField(primary_key=True)
-    user_id = models.DecimalField(max_digits=10, decimal_places=0)
+    user_id = models.IntegerField()
     prod_id = models.ForeignKey(Product, on_delete=models.CASCADE, db_column='prod_id')
-    prod_num = models.DecimalField(max_digits=10, decimal_places=0)
+    prod_num = models.IntegerField()
